backend/requirements.txt
socketio[sqlite]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
spacy
transformers
torch
scikit-learn
opencv-python
pydantic
uvicorn
